집에 가자

+++ project/src/main/java/com/example/model/vo/CustomerVO.java
@@ -11,7 +11,7 @@ public class CustomerVO {
 	private String reg_date;
 	private String social_id;     // 소셜 로그인 고유 ID (api -> social_id)
 	private String provider;      // google / kakao / naver / local
-	private Integer admin_bnum;
+	private Long admin_bnum;
 	private String phone;
 	private String addr;
 	private String email;
